feat: reflect topic attribute properties into EventMetadata, add EventTypeName
- EventMetadata.AttributeProperties captures every public property of the
  discovered topic attribute via reflection (name -> stringified value),
  so custom/unknown attribute properties surface in generated docs even
  though the generator has no dedicated field for them.
- EventMetadata.EventTypeName exposes the CLR type name separately from
  EventName (which may be overridden via the attribute's EventName
  property).
- Extracted EventMetadataBuilder from AssemblyEventDiscovery so assembly
  scanning stays independent from reflecting a single event type into
  EventMetadata.
- Added a GenericAttributeDiscoveryTests case using a custom attribute
  (Domain/Subdomain/empty Topic) to verify the dictionary generalizes
  beyond the well-known EventTopic properties, and that an empty Topic
  still falls back to the kebab-cased type name.
- Added a completeness test (FluidMarkdownGeneratorTests) asserting every
  renderable EventMetadata property actually appears in generated
  markdown, which caught two real gaps fixed alongside this change:
  Domain was computed but never rendered, and TopicAttribute could point
  at the wrong attribute instance when a type had more than one attribute
  (GetCustomAttributes<Attribute>().FirstOrDefault() matches any
  attribute, not specifically the topic one).
