+++ libs/Momentum/src/Momentum.Extensions.EventMarkdownGenerator/Templates/Models/EventViewModel.cs
@@ -5,10 +5,12 @@ namespace Momentum.Extensions.EventMarkdownGenerator.Templates.Models;
 public class EventViewModel
 {
     public string EventName { get; set; } = string.Empty;
+    public string EventTypeName { get; set; } = string.Empty;
     public string FullTypeName { get; set; } = string.Empty;
     public string Namespace { get; set; } = string.Empty;
     public string Topic { get; set; } = string.Empty;
     public string FullyQualifiedTopicName { get; set; } = string.Empty;
+    public string Domain { get; set; } = string.Empty;
     public string Version { get; set; } = string.Empty;
     public string Status { get; set; } = string.Empty;
     public string Entity { get; set; } = string.Empty;
@@ -18,6 +20,7 @@ public class EventViewModel
     public string? ObsoleteMessage { get; set; }
     public string GithubUrl { get; set; } = string.Empty;
     public string TopicAttributeDisplayName { get; set; } = string.Empty;
+    public AttributePropertyViewModel[] AttributeProperties { get; set; } = [];
 
     // Documentation properties
     public string Description { get; set; } = string.Empty;
@@ -29,4 +32,10 @@ public class EventViewModel
     public PartitionKeyViewModel[] PartitionKeys { get; set; } = [];
     public int TotalEstimatedSizeBytes { get; set; }
     public bool HasInaccurateEstimates { get; set; }
+}
+
+public class AttributePropertyViewModel
+{
+    public string Key { get; set; } = string.Empty;
+    public string Value { get; set; } = string.Empty;
 }
