+++ libs/Momentum/tests/Momentum.Extensions.EventMarkdownGenerator.Tests/FluidMarkdownGeneratorTests.cs
@@ -29,6 +29,7 @@ public class FluidMarkdownGeneratorTests
             Metadata = new EventMetadata
             {
                 EventName = "TestEvent",
+                EventTypeName = eventType.Name,
                 FullTypeName = eventType.FullName!,
                 Namespace = eventType.Namespace!,
                 Topic = "tests",
@@ -211,6 +212,108 @@ public class FluidMarkdownGeneratorTests
             // The tail must not spill out as an orphaned line below the table.
             result.Content.Split('\n')
                 .ShouldNotContain(line => line.TrimStart().StartsWith("are delivered in publication order"));
+        }
+        finally
+        {
+            Directory.Delete(outputDir, true);
+        }
+    }
+
+    [Fact]
+    public async Task GenerateMarkdown_ShouldRenderEveryEventMetadataProperty()
+    {
+        var generator = await FluidMarkdownGenerator.CreateAsync();
+
+        var eventType = typeof(TestEvent);
+        var topicAttribute = eventType.GetCustomAttributes(false).First(a => a.GetType().Name.Contains("EventTopic"));
+
+        var eventWithDoc = new EventWithDocumentation
+        {
+            Metadata = new EventMetadata
+            {
+                EventName = "MarkerEventName",
+                EventTypeName = "MarkerEventTypeName",
+                FullTypeName = "Marker.Namespace.MarkerFullTypeName",
+                Namespace = "Marker.Namespace",
+                Topic = "marker-topic",
+                FullyQualifiedTopicName = "{env}.marker-domain.public.marker-topic.v7",
+                Domain = "MarkerDomain",
+                Version = "v7",
+                IsInternal = false,
+                EventType = eventType,
+                TopicAttribute = (Attribute)topicAttribute,
+                AttributeProperties = new Dictionary<string, string>
+                {
+                    ["MarkerAttributeKey"] = "MarkerAttributeValue"
+                },
+                ObsoleteMessage = "MarkerObsoleteMessage",
+                Properties =
+                [
+                    new EventPropertyMetadata
+                    {
+                        Name = "MarkerPropertyName",
+                        TypeName = "MarkerPropertyTypeName",
+                        PropertyType = typeof(string),
+                        IsRequired = true,
+                        IsComplexType = false,
+                        IsPartitionKey = true,
+                        Description = "MarkerPropertyDescription",
+                        EstimatedSizeBytes = 42,
+                        IsAccurate = true
+                    }
+                ],
+                PartitionKeys =
+                [
+                    new PartitionKeyMetadata
+                    {
+                        Name = "MarkerPropertyName",
+                        TypeName = "MarkerPropertyTypeName",
+                        Description = "MarkerPartitionKeyDescription",
+                        Order = 0,
+                        IsFromParameter = false
+                    }
+                ]
+            },
+            Documentation = new EventDocumentation
+            {
+                Summary = "MarkerSummary",
+                Remarks = "MarkerRemarks",
+                Example = "MarkerExample",
+                PropertyDescriptions = new Dictionary<string, string>
+                {
+                    ["MarkerPropertyName"] = "MarkerPropertyDescription"
+                }
+            }
+        };
+
+        var outputDir = CreateTempDirectory();
+
+        try
+        {
+            var result = generator.GenerateMarkdown(eventWithDoc, outputDir);
+
+            // Every scalar EventMetadata property the generator can render has a distinct marker value
+            // above. EventType/TopicAttribute are excluded: they're reflected *sources* (their data
+            // surfaces through EventTypeName/FullTypeName and TopicAttributeDisplayName/AttributeProperties
+            // respectively), not values rendered verbatim themselves.
+            result.Content.ShouldContain("MarkerEventName");
+            result.Content.ShouldContain("MarkerEventTypeName");
+            result.Content.ShouldContain("Marker.Namespace.MarkerFullTypeName");
+            result.Content.ShouldContain("Marker.Namespace");
+            result.Content.ShouldContain("marker-topic");
+            result.Content.ShouldContain("{env}.marker-domain.public.marker-topic.v7");
+            result.Content.ShouldContain("MarkerDomain");
+            result.Content.ShouldContain("v7");
+            result.Content.ShouldContain("MarkerObsoleteMessage");
+            result.Content.ShouldContain("MarkerAttributeKey");
+            result.Content.ShouldContain("MarkerAttributeValue");
+            result.Content.ShouldContain("MarkerPropertyName");
+            result.Content.ShouldContain("MarkerPropertyTypeName");
+            result.Content.ShouldContain("MarkerPropertyDescription");
+            result.Content.ShouldContain("MarkerPartitionKeyDescription");
+            result.Content.ShouldContain("MarkerSummary");
+            result.Content.ShouldContain("MarkerRemarks");
+            result.Content.ShouldContain("MarkerExample");
         }
         finally
         {
