+++ libs/Momentum/tests/Momentum.Extensions.EventMarkdownGenerator.Tests/IntegrationTestScenarios/basic-event/expected/TestEvents.Events.Core.IntegrationEvents.MinimalEvent.md
@@ -8,6 +8,7 @@ editLink: false
 # MinimalEvent
 
 - **Status:** Active
+- **Domain:** TestEvents
 - **Version:** v1
 - **Entity:** `minimal-event`
 - **Type:** Integration Event
@@ -34,5 +35,12 @@ This event uses a partition key for message routing:
     ## Technical Details
 
 - **Full Type:** [TestEvents.Events.Core.IntegrationEvents.MinimalEvent](#)
+- **Type Name:** `MinimalEvent`
 - **Namespace:** `TestEvents.Events.Core.IntegrationEvents`
-- **Topic Attribute:** `[EventTopic]`
+- **Topic Attribute:** `[EventTopic<MinimalEvent>]`
+- **Attribute Properties:**
+- `ShouldPluralizeTopicName`: `True`
+- `Topic`: `minimal-event`
+- `Domain`: *(empty)*
+- `Version`: `v1`
+- `Internal`: `False`
