+++ libs/Momentum/tests/Momentum.Extensions.EventMarkdownGenerator.Tests/IntegrationTestScenarios/circular-references/expected/TestEvents.Graph.Nodes.Contracts.IntegrationEvents.NodeRelationshipCreated.md
@@ -8,6 +8,7 @@ editLink: false
 # NodeRelationshipCreated
 
 - **Status:** Active
+- **Domain:** Nodes
 - **Version:** v1
 - **Entity:** `node-relationship-created`
 - **Type:** Integration Event
@@ -48,5 +49,12 @@ This event uses a partition key for message routing:
 ## Technical Details
 
 - **Full Type:** [TestEvents.Graph.Nodes.Contracts.IntegrationEvents.NodeRelationshipCreated](#)
+- **Type Name:** `NodeRelationshipCreated`
 - **Namespace:** `TestEvents.Graph.Nodes.Contracts.IntegrationEvents`
-- **Topic Attribute:** `[EventTopic]`
+- **Topic Attribute:** `[EventTopic<NodeRelationshipCreated>]`
+- **Attribute Properties:**
+- `ShouldPluralizeTopicName`: `True`
+- `Topic`: `node-relationship-created`
+- `Domain`: *(empty)*
+- `Version`: `v1`
+- `Internal`: `False`
