+++ libs/Momentum/tests/Momentum.Extensions.EventMarkdownGenerator.Tests/IntegrationTestScenarios/collection-types/expected/TestEvents.AppDomain.Collections.Contracts.IntegrationEvents.BulkDataProcessed.md
@@ -8,6 +8,7 @@ editLink: false
 # BulkDataProcessed
 
 - **Status:** Active
+- **Domain:** Collections
 - **Version:** v1
 - **Entity:** `bulk-data-processed`
 - **Type:** Integration Event
@@ -63,5 +64,12 @@ This event uses a partition key for message routing:
 ## Technical Details
 
 - **Full Type:** [TestEvents.AppDomain.Collections.Contracts.IntegrationEvents.BulkDataProcessed](#)
+- **Type Name:** `BulkDataProcessed`
 - **Namespace:** `TestEvents.AppDomain.Collections.Contracts.IntegrationEvents`
-- **Topic Attribute:** `[EventTopic]`
+- **Topic Attribute:** `[EventTopic<BulkDataProcessed>]`
+- **Attribute Properties:**
+- `ShouldPluralizeTopicName`: `True`
+- `Topic`: `bulk-data-processed`
+- `Domain`: *(empty)*
+- `Version`: `v1`
+- `Internal`: `False`
