+++ libs/Momentum/tests/Momentum.Extensions.EventMarkdownGenerator.Tests/IntegrationTestScenarios/complex-event-with-schemas/expected/TestEvents.AppDomain.Orders.Contracts.IntegrationEvents.OrderCompleted.md
@@ -8,6 +8,7 @@ editLink: false
 # OrderCompleted
 
 - **Status:** Active
+- **Domain:** Orders
 - **Version:** v1
 - **Entity:** `order`
 - **Type:** Integration Event
@@ -51,5 +52,12 @@ This event uses a partition key for message routing:
 ## Technical Details
 
 - **Full Type:** [TestEvents.AppDomain.Orders.Contracts.IntegrationEvents.OrderCompleted](#)
+- **Type Name:** `OrderCompleted`
 - **Namespace:** `TestEvents.AppDomain.Orders.Contracts.IntegrationEvents`
-- **Topic Attribute:** `[EventTopic]`
+- **Topic Attribute:** `[EventTopic<Order>]`
+- **Attribute Properties:**
+- `ShouldPluralizeTopicName`: `True`
+- `Topic`: `order`
+- `Domain`: *(empty)*
+- `Version`: `v1`
+- `Internal`: `False`
