+++ libs/Momentum/tests/Momentum.Extensions.EventMarkdownGenerator.Tests/IntegrationTestScenarios/missing-documentation/expected/TestEvents.Missing.Docs.IntegrationEvents.MalformedXmlEvent.md
@@ -8,6 +8,7 @@ editLink: false
 # MalformedXmlEvent
 
 - **Status:** Active
+- **Domain:** TestEvents
 - **Version:** v1
 - **Entity:** `malformed-xml-event`
 - **Type:** Integration Event
@@ -35,5 +36,12 @@ This event uses a partition key for message routing:
     ## Technical Details
 
 - **Full Type:** [TestEvents.Missing.Docs.IntegrationEvents.MalformedXmlEvent](#)
+- **Type Name:** `MalformedXmlEvent`
 - **Namespace:** `TestEvents.Missing.Docs.IntegrationEvents`
-- **Topic Attribute:** `[EventTopic]`
+- **Topic Attribute:** `[EventTopic<MalformedXmlEvent>]`
+- **Attribute Properties:**
+- `ShouldPluralizeTopicName`: `True`
+- `Topic`: `malformed-xml-event`
+- `Domain`: *(empty)*
+- `Version`: `v1`
+- `Internal`: `False`
