+++ libs/Momentum/tests/Momentum.Extensions.EventMarkdownGenerator.Tests/IntegrationTestScenarios/multiple-events-sidebar/expected/TestEvents.AppDomain.Legacy.Contracts.IntegrationEvents.PaymentProcessed.md
@@ -11,6 +11,7 @@ editLink: false
     > This event is deprecated. This event has been replaced by the new PaymentCompleted event. Please migrate to use the new event structure for enhanced functionality.
 
 - **Status:** Deprecated
+- **Domain:** Legacy
 - **Version:** v1
 - **Entity:** `payment-processed`
 - **Type:** Integration Event
@@ -39,5 +40,12 @@ This event uses a partition key for message routing:
     ## Technical Details
 
 - **Full Type:** [TestEvents.AppDomain.Legacy.Contracts.IntegrationEvents.PaymentProcessed](#)
+- **Type Name:** `PaymentProcessed`
 - **Namespace:** `TestEvents.AppDomain.Legacy.Contracts.IntegrationEvents`
-- **Topic Attribute:** `[EventTopic]`
+- **Topic Attribute:** `[EventTopic<PaymentProcessed>]`
+- **Attribute Properties:**
+- `ShouldPluralizeTopicName`: `True`
+- `Topic`: `payment-processed`
+- `Domain`: *(empty)*
+- `Version`: `v1`
+- `Internal`: `False`
