+++ libs/Momentum/tests/Momentum.Extensions.EventMarkdownGenerator.Tests/IntegrationTestScenarios/multiple-partition-keys/expected/TestEvents.AppDomain.Inventory.Contracts.IntegrationEvents.ProductStockUpdated.md
@@ -8,6 +8,7 @@ editLink: false
 # ProductStockUpdated
 
 - **Status:** Active
+- **Domain:** Inventory
 - **Version:** v1
 - **Entity:** `product`
 - **Type:** Integration Event
@@ -42,5 +43,12 @@ This event uses multiple partition keys for message routing:
     ## Technical Details
 
 - **Full Type:** [TestEvents.AppDomain.Inventory.Contracts.IntegrationEvents.ProductStockUpdated](#)
+- **Type Name:** `ProductStockUpdated`
 - **Namespace:** `TestEvents.AppDomain.Inventory.Contracts.IntegrationEvents`
-- **Topic Attribute:** `[EventTopic]`
+- **Topic Attribute:** `[EventTopic<Product>]`
+- **Attribute Properties:**
+- `ShouldPluralizeTopicName`: `True`
+- `Topic`: `product`
+- `Domain`: *(empty)*
+- `Version`: `v1`
+- `Internal`: `False`
